# TBS: perf: Pool APMEvent in sampling decision handler