--- a/x-pack/apm-server/sampling/eventstorage/storage.go
+++ b/x-pack/apm-server/sampling/eventstorage/storage.go
@@ -288,16 +288,16 @@ func (rw *ReadWriter) ReadTraceEvents(traceID string, out *modelpb.Batch) error 
 		}
 		switch item.UserMeta() {
 		case entryMetaTraceEvent:
-			var event modelpb.APMEvent
+			event := modelpb.APMEventFromVTPool()
 			if err := item.Value(func(data []byte) error {
-				if err := rw.s.codec.DecodeEvent(data, &event); err != nil {
+				if err := rw.s.codec.DecodeEvent(data, event); err != nil {
 					return fmt.Errorf("codec failed to decode event: %w", err)
 				}
 				return nil
 			}); err != nil {
 				return err
 			}
-			*out = append(*out, &event)
+			*out = append(*out, event)
 		default:
 			// Unknown entry meta: ignore.
 			continue