--- a/x-pack/apm-server/sampling/processor_test.go
+++ b/x-pack/apm-server/sampling/processor_test.go
@@ -429,10 +429,14 @@ func TestProcessRemoteTailSampling(t *testing.T) {
 
 	reported := make(chan modelpb.Batch)
 	config.BatchProcessor = modelpb.ProcessBatchFunc(func(ctx context.Context, batch *modelpb.Batch) error {
+		batchCopy := make(modelpb.Batch, len(*batch))
+		for i := range *batch {
+			batchCopy[i] = (*batch)[i].CloneVT()
+		}
 		select {
 		case <-ctx.Done():
 			return ctx.Err()
-		case reported <- *batch:
+		case reported <- batchCopy:
 			return nil
 		}
 	})